Simplify handling of new-window events
